moved abstract class from control.py
